Trying...

+++ BasiliskII/src/Unix/sysdeps.h
@@ -27,7 +27,7 @@
 #error "Your compiler is not ANSI. Get a real one."
 #endif
 
-#include "config.h"
+#include <config.h>
 #include "user_strings_unix.h"
 
 #ifndef STDC_HEADERS
@@ -221,7 +221,7 @@ typedef uae_u32 uaecptr;
 
 /* Timing functions */
 extern uint64 GetTicks_usec(void);
-extern void Delay_usec(uint32 usec);
+extern void Delay_usec(uint64 usec);
 
 /* Spinlocks */
 #ifdef __GNUC__
